backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
psycopg2-binary==2.9.6
langchain==0.0.282
google-gemini-api==0.1.0
python-dotenv==1.0.0
pandas==2.0.3
